Add simple wav filter

+++ Compressor.cpp
@@ -64,7 +64,7 @@ void BitStreamCompressor::decompressBytes(byte* in, byte* out, size_t count) {
 }
 
 void Store::compress(Stream* in, Stream* out, uint64_t count) {
-	static const size_t kBufferSize = 8 * KB;
+	static const uint64_t kBufferSize = 8 * KB;
 	byte buffer[kBufferSize];
 	while (count > 0) {
 		const size_t read = in->read(buffer, std::min(count, kBufferSize));
@@ -77,7 +77,7 @@ void Store::compress(Stream* in, Stream* out, uint64_t count) {
 }
 
 void Store::decompress(Stream* in, Stream* out, uint64_t count) {
-	static const size_t kBufferSize = 8 * KB;
+	static const uint64_t kBufferSize = 8 * KB;
 	byte buffer[kBufferSize];
 	while (count > 0) {
 		const size_t read = in->read(buffer, std::min(count, kBufferSize));
